Fix audio buffer data type

+++ src/apu/audio_stream.rs
@@ -6,12 +6,12 @@ use rodio::buffer::SamplesBuffer;
 use rodio::Sink;
 
 pub struct AudioStream {
-  tx: Sender<Vec<f32>>,
+  tx: Sender<Vec<i16>>,
 }
 
 impl AudioStream {
   pub fn new() -> AudioStream {
-    let (tx, rx): (Sender<Vec<f32>>, Receiver<Vec<f32>>) = mpsc::channel();
+    let (tx, rx): (Sender<Vec<i16>>, Receiver<Vec<i16>>) = mpsc::channel();
 
     thread::spawn(move || {
       let (_stream, stream_handle) = rodio::OutputStream::try_default().unwrap();
@@ -29,7 +29,7 @@ impl AudioStream {
     }
   }
 
-  pub fn send_audio_buffer(&mut self, sample: Vec<f32>) {
+  pub fn send_audio_buffer(&mut self, sample: Vec<i16>) {
     let _ = self.tx.send(sample);
   }
 }
